SetStatusDescriptor: extract to its own file
